requirements.txt
ansicon==1.89.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
blessed==1.18.0
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
filelock==3.0.12
flyingcircus==0.1.3.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
jinxed==1.0.1
mysql-connector-python==8.0.23
Pillow==8.0.1
psycopg2==2.8.6
pyTelegramBotAPI==3.7.6
python-telegram-bot==13.3
pytk==0.0.2.1
pytz==2020.4
requests==2.25.1
setuptools-scm==5.0.2
six==1.15.0
sqlparse==0.4.1
telebot==0.0.4
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
virtualenv==20.1.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.0
